refactory: now both Troop and Tower extend Entity

+++ src/main/java/jroyale/shared/Side.java
@@ -3,6 +3,6 @@ package jroyale.shared;
 public final class Side {
     Side() {} // empty constructor
 
-    public static final int PLAYER = 0;
-    public static final int OPPONENT = 1;
+    public static final byte PLAYER = 0;
+    public static final byte OPPONENT = 1;
 }
