+++ src/main/java/jroyale/shared/TowerIndex.java
@@ -4,11 +4,11 @@ public final class TowerIndex {
     TowerIndex() {} // empty constructor
 
     // towers index that are visible everywhere
-    public static final int PLAYER_KING_TOWER = 0;
-    public static final int PLAYER_LEFT_TOWER = 1;
-    public static final int PLAYER_RIGHT_TOWER = 2;
-    public static final int OPPONENT_KING_TOWER = 3;
-    public static final int OPPONENT_LEFT_TOWER = 4;
-    public static final int OPPONENT_RIGHT_TOWER = 5;
-    public static final int NUM_TOWERS = 6;
+    public static final byte PLAYER_KING_TOWER = 0;
+    public static final byte PLAYER_LEFT_TOWER = 1;
+    public static final byte PLAYER_RIGHT_TOWER = 2;
+    public static final byte OPPONENT_KING_TOWER = 3;
+    public static final byte OPPONENT_LEFT_TOWER = 4;
+    public static final byte OPPONENT_RIGHT_TOWER = 5;
+    public static final byte NUM_TOWERS = 6;
 }
